# openapi.yaml
openapi: 3.1.0
info:
  title: Meeting Management API
  version: 1.0.0
  description: MVP API for meeting scheduling, rooms, participants, attachments, email invitations, RSVP, and reminders.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Sign in with an existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 8 }
      responses:
        '200':
          description: Authenticated; access token expires in 30 minutes
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [access_token, token_type, expires_in, user]
                        properties:
                          access_token: { type: string }
                          token_type: { type: string, const: Bearer }
                          expires_in: { type: integer, const: 1800 }
                          user: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      tags: [Auth]
      summary: Return the authenticated user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Meta:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        requestId: { type: string }
    Success:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, const: true }
        data: {}
        meta: { $ref: '#/components/schemas/Meta' }
    Error:
      type: object
      required: [success, error, meta]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              description: For validation errors each item includes field, code, and localized message; conflict items identify the conflicting room/user/meeting.
              items:
                type: object
                properties:
                  field: { type: string }
                  code: { type: string }
                  message: { type: string }
                  entity_id: { type: integer }
                  entity_name: { type: string }
                  meeting_id: { type: integer }
            suggested_times:
              type: array
              maxItems: 3
              items:
                type: object
                required: [start_time, end_time]
                properties:
                  start_time: { type: string, format: date-time }
                  end_time: { type: string, format: date-time }
            override_allowed: { type: boolean }
        meta: { $ref: '#/components/schemas/Meta' }
...
    User:
      type: object
      required: [user_id, full_name, email, role]
      properties:
        user_id: { type: integer }
        full_name: { type: string }
        email: { type: string, format: email }
        phone: { type: [string, 'null'] }
        role: { type: string, enum: [Admin, Organizer, Participant] }
...
    ErrorResponse:
      allOf: [{ $ref: '#/components/schemas/Error' }]
  responses:
    ValidationError:
      description: Invalid request fields or date range. Field messages include “Vui lòng nhập tên cuộc họp”, “Thời gian kết thúc phải sau thời gian bắt đầu”, and “Vui lòng chọn ít nhất 1 người tham gia”.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing, invalid, or expired access token / invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...